mv BuildNativeInvokeCode from http to transactionbuilder
